Added the frontend and edited some files

+++ server/controllers/customer.js
@@ -2,7 +2,7 @@ import Customer from '../models/customer.js';
 import jwt from 'jsonwebtoken';
 import Product from '../models/product.js';
 import Order from '../models/order.js'
-import { customerSchema, addressSchema, customerEditSchema } from '../joiSchemas.js';
+import { customerSchema, addressSchema, customerEditSchema, orderSchema } from '../joiSchemas.js';
 import Address from '../models/address.js';
 
 
@@ -12,7 +12,7 @@ export const register = async (req, res) => {
         if (error) {
             return res.status(400).json({ error: error.details[0].message });
         }
-        const { name, username, email, password, phone } = req.body;
+        const { name, username, email, password, phone, gender } = req.body;
 
         // Check if the username already exists
         const existingUser = await Customer.findOne({ username });
@@ -33,7 +33,7 @@ export const register = async (req, res) => {
         }
 
         // Use passport-local-mongoose to register the user
-        const user = await Customer.register(new Customer({ name, username, email, phone }), password);
+        const user = await Customer.register(new Customer({ name, username, email, phone, gender }), password);
 
         res.status(201).json({ message: 'User registered successfully' });
     } catch (error) {
@@ -49,15 +49,15 @@ export const login = async (req, res, next, secret) => {
             $or: [{ username: identifier }, { email: identifier }]
         });
         if (!user) {
-            return res.status(400).json({ error: 'Invalid username or email' });
+            return res.status(400).json('Invalid username or email');
         }
 
         user.authenticate(password, (err, authenticatedUser, error) => {
             if (err) { return next(err); }
-            if (!authenticatedUser) { return res.status(400).json({ error: 'Invalid username or email' }); }
+            if (!authenticatedUser) { return res.status(400).json('Invalid username or email'); }
             const token = jwt.sign({ sub: authenticatedUser._id, userType: 'customer' }, secret, { expiresIn: '1hr' });
             res.cookie('acc_token', token, { httpOnly: true })
-            res.json(token);
+            res.json({ token });
         });
     } catch (error) {
         res.status(500).json("Something went wrong!");
@@ -71,7 +71,7 @@ export const home = async (req, res, next) => {
     const skip = (page - 1) * limit;
 
     try {
-        const products = await Product.find({ isAvailable: true }).skip(skip).limit(limit);
+        const products = await Product.find({ isAvailable: true }).skip(skip).limit(limit).populate("seller");
         res.json(products);
     } catch (error) {
         res.status(500).json({ error: error.message });
@@ -80,12 +80,38 @@ export const home = async (req, res, next) => {
 
 export const viewProduct = async (req, res, next) => {
     try {
-        const { id } = req.params;
-        const product = await Product.findById(id);
+        const { productId } = req.params;
+        const product = await Product.findById(productId);
         if (!product) {
             return res.status(404).json({ error: 'Product not found' });
         }
-        res.json(product);
+        res.json({ product });
+    } catch (error) {
+        res.status(500).json({ error: "Something went wrong" });
+    }
+}
+export const showAllCategories = async (req, res) => {
+    try {
+        const categories = await Product.aggregate([
+            { $match: { isAvailable: true } },
+            { $group: { _id: "$category" } }
+        ]);
+        const availableCategories = categories.map(category => category._id);
+        res.json(availableCategories);
+    } catch (error) {
+        res.status(500).json({ error: "Something went wrong" });
+    }
+}
+
+
+export const categorizeProducts = async (req, res) => {
+    try {
+        const page = parseInt(req.query.page) || 1;
+        const limit = parseInt(req.query.limit) || 10;
+        const skip = (page - 1) * limit;
+        const { category } = req.query;
+        const filteredProducts = await Product.find({ category, isAvailable: true }).skip(skip).limit(limit);
+        res.json(filteredProducts);
     } catch (error) {
         res.status(500).json({ error: "Something went wrong" });
     }
@@ -115,6 +141,9 @@ export const addToCart = async (req, res, next) => {
             if (customer.cart[cartProductIndex].quantity + 1 > product.quantity) {
                 return res.status(400).json({ error: 'Not enough product in stock' });
             }
+            if (customer.cart[cartProductIndex].quantity >= 5) {
+                return res.status(400).json({ error: "We're sorry! Only 5 unit(s) allowed in each orders" });
+            }
             customer.cart[cartProductIndex].quantity += 1;
             customer.cart[cartProductIndex].price += product.price; // Assuming product.price is the price per item
         } else {
@@ -131,22 +160,36 @@ export const addToCart = async (req, res, next) => {
 
 
 export const searchProduct = async (req, res) => {
+    const page = parseInt(req.query.page) || 1;
+    const limit = parseInt(req.query.limit) || 10;
+    const skip = (page - 1) * limit;
+    if (!req.query.q) {
+        const products = await Product.find().skip(skip).limit(limit).lean();
+        return res.json(products)
+    }
     try {
-        const query = req.query.q;
-        const products = await Product.find({ name: new RegExp(query, 'i') });
-        if (products.length === 0) {
+        let query = req.query.q;
+        if (query.length <= 2) {
             return res.status(404).json({ message: 'No products found matching your search criteria' });
         }
-        res.json(products);
+        query = query.replace(/\s/g, ''); // Remove spaces from the query
+        const regex = new RegExp(query.split("").join("\\s*"), 'i');
+        const matchedProducts = await Product.find({ productName: { $regex: regex }, isAvailable: true }).skip(skip).limit(limit).lean();
+        if (matchedProducts.length === 0) {
+            return res.status(404).json({ message: 'No products found matching your search criteria' });
+        }
+        res.json(matchedProducts);
     } catch (error) {
         console.error(error); // Log the error
         res.status(500).json({ error: 'An error occurred while searching for the product.' });
     }
 }
 
+
+
 export const updateCart = async (req, res, next) => {
     try {
-        const { productId, action } = req.params; // action can be 'decrease' or 'remove'
+        const { productId, action } = req.params; // action can be 'increase', 'decrease', or 'remove'
         const customerId = req.user._id;
         const customer = await Customer.findById(customerId);
         if (!customer) {
@@ -156,11 +199,25 @@ export const updateCart = async (req, res, next) => {
         // Find the product in the cart
         const cartProductIndex = customer.cart.findIndex(cp => cp.product.toString() === productId);
 
-        // If the product is in the cart, decrease the quantity or remove it
+        // If the product is in the cart, increase or decrease the quantity, or remove it
         if (cartProductIndex >= 0) {
-            if (action === 'decrease' && customer.cart[cartProductIndex].quantity > 1) {
+            const product = await Product.findById(productId); // Fetch the product to get its price
+            if (!product) {
+                return res.status(404).json({ error: 'Product not found' });
+            }
+
+            if (action === 'increase') {
+                if (customer.cart[cartProductIndex].quantity >= 5) {
+                    return res.status(400).json({ error: "We're sorry! Only 5 unit(s) allowed in each orders" });
+                }
+                customer.cart[cartProductIndex].quantity += 1;
+                customer.cart[cartProductIndex].price += product.price;
+            } else if (action === 'decrease') {
+                if (customer.cart[cartProductIndex].quantity <= 1) {
+                    return res.status(400).json({ error: 'Minimum quantity: 1' });
+                }
                 customer.cart[cartProductIndex].quantity -= 1;
-                customer.cart[cartProductIndex].price -= customer.cart[cartProductIndex].price / (customer.cart[cartProductIndex].quantity + 1);
+                customer.cart[cartProductIndex].price -= product.price;
             } else if (action === 'remove') {
                 customer.cart.splice(cartProductIndex, 1);
             } else {
@@ -175,16 +232,55 @@ export const updateCart = async (req, res, next) => {
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
-}
+};
+
 
 
 export const showCart = async (req, res, next) => {
     try {
         const customerId = req.user._id;
+        let grandTotal = 0
         const customer = await Customer.findById(customerId).populate("cart.product");
+        const userDetails = { name: customer.name, email: customer.email, phone: customer.phone }
         const products = customer.cart;
+        for (let product of products) {
+            grandTotal += product.price
+        }
+        customer.cart.length ? res.json({ products, grandTotal, userDetails }) : res.json("Your cart is empty");
+    } catch (error) {
+        res.status(500).json({ error: error.message });
+    }
+}
+export const checkout = async (req, res, next) => {
+    try {
+        const customerId = req.user._id;
+        const customer = await Customer.findById(customerId).populate(["cart.product", "address"]);
+        if (customer.address.length > 0) {
+            const userDetails = { name: customer.name, email: customer.email, phone: customer.phone, address: customer.address }
+            if (req.body.productId) {
+                const { productId } = req.body
+                const product = await Product.findById(productId)
+                const grandTotal = product.price;
+                return res.json({ product, grandTotal, userDetails })
+            } else {
+                for (let cartProduct of customer.cart) {
+                    const product = await Product.findById(cartProduct.product._id);
+                    const quantity = cartProduct.quantity;
+                    if (product.quantity < quantity) {
+                        return res.status(400).json({ error: `Uh-oh! Only ${product.quantity} units of ${product.productName} are left in stock.` })
+                    }
+                }
+                let grandTotal = 0
+                const products = customer.cart;
 
-        customer.cart.length ? res.json({ products }) : res.json("Your cart is empty");
+                for (let product of products) {
+                    grandTotal += product.price
+                }
+                customer.cart.length ? res.json({ grandTotal, userDetails }) : res.json("Your cart is empty");
+            }
+        } else {
+            return res.status(400).json({ error: "Add your Address" })
+        }
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
@@ -221,13 +317,19 @@ export const orderProduct = async (req, res, next) => {
             } else {
                 // Buy All Products in Cart
                 if (customer.cart.length) {
+
                     await customer.populate('cart.product');
+
+                    for (let cartProduct of customer.cart) {
+                        const product = await Product.findById(cartProduct.product._id);
+                        const quantity = cartProduct.quantity;
+                        if (product.quantity < quantity) {
+                            return res.status(400).json(`Uh-oh! Only ${product.quantity} units of ${product.productName} are left in stock.`)
+                        }
+                    }
                     for (let cartProduct of customer.cart) {
                         const product = await Product.findById(cartProduct.product._id);
                         const quantity = cartProduct.quantity; // Get the quantity from the cart
-                        if (product.quantity < quantity) {
-                            return res.json("Out of Stock")
-                        }
                         const total = product.price * quantity; // Multiply the product price by the quantity
                         orderProducts.push({ product, quantity, total });
                         grandTotal += total;
@@ -236,16 +338,30 @@ export const orderProduct = async (req, res, next) => {
                         product.quantity -= quantity;
                         await product.save();
                     }
+
                 } else {
                     return res.status(404).json({ error: 'Your cart is empty' });
                 }
             }
-            // Create order with products
+            const { error } = orderSchema.validate(req.body);
+            if (error) {
+                return res.status(400).json({ error: error.details[0].message });
+            }
+            const { payedPrice, paymentMethod, address } = req.body;
             const order = new Order({
                 customer,
                 products: orderProducts,
-                grandTotal,
-                shippingAddress: customer.address[0] // Assuming the first address is the shipping address
+                payment: { totalPrice: grandTotal, payedPrice, paymentMethod },
+                shippingAddress: {
+                    addressType: address.addressType,
+                    buildingName: address.buildingName,
+                    street: address.street,
+                    district: address.district,
+                    landmark: address.landmark,
+                    state: address.state,
+                    pincode: address.pincode,
+                    phone: address.phone
+                }
             });
             await order.save();
             // Add the order to the customer's order history
@@ -258,7 +374,7 @@ export const orderProduct = async (req, res, next) => {
             }
             res.json({ message: 'Order placed successfully!' });
         } else {
-            res.json({ message: "Add your Address" })
+            res.status(400).json({ error: "Add your Address" })
         }
     } catch (error) {
         console.error(error); // Log the error
@@ -271,18 +387,27 @@ export const orderHistory = async (req, res) => {
         const customer = await Customer.findById(req.user._id)
             .populate({
                 path: 'orderHistory',
+                options: { sort: { 'createdAt': -1 } }, // Sort orders by createdAt in descending order
                 populate: {
                     path: 'products.product', // populate product in each product object
-                },
-                populate: {
-                    path: "shippingAddress"
                 }
             });
         const allOrders = customer.orderHistory;
         if (!allOrders.length) {
             return res.json("Start shopping to fill your order history!")
         }
-        res.status(200).json(allOrders);
+        const allOrderedProducts = []
+        const otherDetails = []
+        for (let order of allOrders) {
+            allOrderedProducts.push(order.products)
+            const createdAt = new Date(order.createdAt);
+            const options = { year: 'numeric', month: 'long', day: 'numeric', hour: '2-digit', minute: '2-digit', second: '2-digit' };
+
+            const formattedDate = `Ordered on ${createdAt.toLocaleDateString(undefined, options)}`;
+            otherDetails.push([order.payment, order.shippingAddress, formattedDate, order._id])
+        }
+
+        res.status(200).json({ otherDetails, allOrderedProducts });
     } catch (error) {
         console.log(error)
         res.status(500).json({ error: 'An error occurred while retrieving the order history' });
@@ -325,7 +450,7 @@ export const editProfile = async (req, res) => {
         await customer.save();
 
         // Send the updated customer data as a response
-        res.status(200).json(customer);
+        res.status(200).json({ message: "User Details edited successfully" });
     } catch (error) {
         // Handle any errors
         res.status(500).json({ error: 'An error occurred while updating the profile.' });
@@ -355,7 +480,7 @@ export const addAddress = async (req, res) => {
         // Save the updated customer data
         await customer.save();
         // Send the updated customer data as a response
-        res.status(200).json(customer);
+        res.status(200).json({ message: "Address added successfully" });
     }
     catch (error) {
         // Handle any errors
@@ -364,7 +489,15 @@ export const addAddress = async (req, res) => {
     }
 
 }
-
+export const viewAddress = async (req, res) => {
+    try {
+        const { addressId } = req.params;
+        const address = await Address.findById(addressId)
+        res.json(address)
+    } catch (error) {
+        res.status(500).json({ error: 'An error occurred while retrieving the address.' });
+    }
+}
 export const editAddress = async (req, res) => {
     try {
         const { addressId } = req.params;
@@ -394,20 +527,23 @@ export const editAddress = async (req, res) => {
 export const deleteAddress = async (req, res) => {
     const { addressId } = req.params;
     try {
-        await Address.findByIdAndDelete(addressId);
-
         // Find the customer and remove the address reference
         const customerId = req.user._id;
         const customer = await Customer.findById(customerId);
-        customer.address = customer.address.filter(id => id.toString() !== addressId);
-        await customer.save();
-
-        res.json("Address removed successfully");
+        if (customer.address.length > 1) {
+            await Address.findByIdAndDelete(addressId);
+            customer.address = customer.address.filter(id => id.toString() !== addressId);
+            await customer.save();
+            res.json({ message: "Address removed successfully" });
+        } else {
+            res.status(400).json({ error: 'Cannot remove the last remaining address.' });
+        }
     }
     catch (error) {
         // Handle any errors
         res.status(500).json({ error: 'An error occurred while removing the address.' });
     }
 }
+
